SONARIAC-805 [Microsoft.Blockchain/blockchainMembers] resources[]/properties/firewallRules/startIpAddress and endIpAddress
Refactoring of the code, before implementation
